refactored elastic net code
